Refactor ddp.py to make measure() reusable
